notmuch: show additional mail in query windows
Sometimes a user may wish to include certain messages in a query window
even if it exists outside the date range. For example, always showing
flagged messages.

To support this, extend the query window code to emit a window
containing an `OR` statement with additional search terms if a user
configures the `nm_query_window_or_terms` variable.

For example, using the mailbox `notmuch://?query=tag:inbox` as the
mailbox and `tag:flagged and tag:unread` as the or terms will produce a
query window such as:

`notmuch://?query=tag:inbox and (date:... or (tag:flagged and tag:unread))`

The `date` value was omitted for brevity. The resulting mailbox contains
mail with tag inbox within the date range and mail with tags inbox,
flagged, and unread.

Closes #2972