config: Add setting to skip validators during startup
An upcoming patch wants to add a circular dependency between the
validators for a pair of config variables (one of type sort, the other
of type enum), so that no matter which variable is set second, the
user can't force an inconsistent setting between the two.  However, it
is impossible to query the value of a variable that has not yet been
registered, and since the dependency will be circular, there is no way
to rearrange the order in which the two variables are registered such
that the validator for each can blindly assume the other is registered
in order to query its value.

We need a way to skip the validators during the initial registration,
and similarly during 'reset all'.

The least-invasive way is to add a flag to the config set that tracks
when we are resetting a variable as part of its initial registration, at
which point there is no need to run a validator, since the compile-time
default value is under our control and therefore valid.  A couple of
unit tests have to mess with the flag as well, in order to properly
simulate when validators fire.

A more direct approach would entail adding a new 'bool init' parameter
to the ConfigSetType::reset() interface instead of changing the config
set, but that would impact more of the code base.  It might also be
possible to temporarily hack cdef->validator to NULL during the initial
registration process, but that feels dirty.