threads: Add and document $use_threads
For now, the variable just makes the decision between threaded and
flat views easier to reason about.  By making the new variable take an
enum of 4 values, but only documenting 3, we can easily tell whether
the user has ever touched it (opting in to our enhanced semantics), or
if they use a legacy config file (in which case we don't do anything
different from before).  And in the short term, when threads are
enabled, the threads are (still) sorted solely by $sort_aux.

This patch takes care of all interactions between $use_threads and
$sort: updating either variable with 'set' causes an observer to
update the other as needed (this is the first time an observer has
modified a ConfigSubset, which must be done carefully to avoid
infinite recursion).  The <sort-mailbox> and <sort-reverse> commands
are made a bit smarter (once opting into the new mode, 'ot' becomes a
way to toggle threads on or off, and all other sort categories
preserve the current threading state).

It also adds some much-needed documentation with worked examples, as
part of a new feature page.

Later patches will then further enhance it so that use_threads=yes
sort=last-date sort_aux=date sorts in the same way as Thunderbird, as
well as adding a %T expando in status_format for better display of the
current sorting state.

Possible improvements:
- the warning produced for interactive ':set sort=threads
use_threads=threads' flashes by too quickly to read (it gets replaced
by 'Sorting mailbox...' in short order); it can be seen later in
<show-log-messages>, but we may want to enhance the warning logic to
pause after certain warnings
- the behavior of use_threads as an enum is not quite the same as
quad-options, which are special-cased in parse_set() such that 'set
quit' is the same as 'set quit=yes' instead of 'set ?quit'.  It would
be nice to add a new DT_ENUM_YES_NO that allows random enums to choose
their own bare set and toggle behavior (in fact, quad-option could
then be just another ordinary enum, for less special-casing elsewhere)