threads: Use $sort over $sort_aux when $use_threads=yes
Fixes: https://github.com/neomutt/neomutt/issues/2342

Now that we have $use_threads, we can use $sort as the specification
of how to sort overall threads, orthogonal to $sort_aux for how to
sort within a thread.  This allows us to finally achieve parity with
Thunderbird, where a thread with the most recent activity (at any
depth) is floated to the top or bottom of the list
(sort_thread=reverse-last-date or last-date), but within the thread,
the subthreads are sorted by the date that sibling subthreads started
and not by when out-of-order replies to those subthreads were added
(sort_aux=date).

Implementation-wise, this requires adding support for parsing 'last-'
as part of $sort.  For convenience, the <sort-mailbox> command does
NOT change the setting of the 'last-' flag.  Also, each thread now has
to track two sort keys: the message to use when sorting intra-thread
siblings, and the message to use when sorting between two overall
threads.  While we are still able to implement sort_aux=last-* with an
O(1) peek at the first or last sibling, properly implementing
sort_thread=last-* requires an O(n) pass over all siblings (although
our runtime is still dominated by the O(n lg n) nature of sorting in
general).